Genetic playout tuning improvements.

+++ FastBoard.h
@@ -55,7 +55,7 @@ public:
         move generation types
     */
     typedef std::tr1::array<int, 24> movelist_t;
-    typedef std::pair<int, int> movescore_t;
+    typedef std::pair<int, float> movescore_t;
     typedef std::tr1::array<movescore_t, 24> scoredlist_t;
     
     int get_boardsize(void);
